Additional Python 3 refactoring.
Changed the way hashlib is imported and added appropriate steps to create the SHA1 digest.

Used an OrderedDict in the creation of the WSSE authentication header in order to guarantee the proper ordering of the header variables.

Converted base64 encoded byte strings into unicode strings so the PasswordDigests and Nonce values work properly in the REST call.